functions/requirements.txt
azure-functions>=1.20
azure-identity>=1.16
httpx>=0.27
